[ACC-1007] Add driver that will pull metrics on a schedule

+++ surveyor-spi/src/main/java/com/contentgrid/surveyor/spi/storage/AggregateEventCountMetricSpiPort.java
@@ -1,15 +1,20 @@
 package com.contentgrid.surveyor.spi.storage;
 
-import java.math.BigInteger;
+import com.contentgrid.surveyor.spi.TimeInterval;
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.util.List;
 
 public interface AggregateEventCountMetricSpiPort {
+
     default EventCountMetric getAggregatedEventCountMetric(Resource resource, TimeInterval interval) {
         return Util.onlyValue(
-                getAggregatedEventCountMetrics(resource, interval, Duration.between(interval.getStartTime(), interval.getEndTime())),
-                () -> new EventCountMetric(interval, resource, BigInteger.ZERO)
+                getAggregatedEventCountMetrics(resource, interval,
+                        Duration.between(interval.getStartTime(), interval.getEndTime())),
+                () -> new EventCountMetric(interval, resource, BigDecimal.ZERO)
         );
     }
-    List<EventCountMetric> getAggregatedEventCountMetrics(Resource resource, TimeInterval interval, Duration chunkDuration);
+
+    List<EventCountMetric> getAggregatedEventCountMetrics(Resource resource, TimeInterval interval,
+            Duration chunkDuration);
 }
